@@ -32,6 +32,7 @@ setup(name='katsdptelstate',
       author='MeerKAT SDP team',
       author_email='sdpdev+katsdptelstate@ska.ac.za',
       packages=find_packages(),
+      package_data={'': ['lua_scripts/*.lua']},
       url='https://github.com/ska-sa/katsdptelstate',
       license='Modified BSD',
       classifiers=[
@@ -40,22 +41,18 @@ setup(name='katsdptelstate',
           'License :: OSI Approved :: BSD License',
           'Operating System :: OS Independent',
           'Programming Language :: Python',
-          'Programming Language :: Python :: 2',
-          'Programming Language :: Python :: 2.7',
-          'Programming Language :: Python :: 3',
-          'Programming Language :: Python :: 3.3',
-          'Programming Language :: Python :: 3.4',
           'Programming Language :: Python :: 3.5',
           'Programming Language :: Python :: 3.6',
           'Programming Language :: Python :: 3.7',
+          'Programming Language :: Python :: 3.8',
           'Topic :: Software Development :: Libraries :: Python Modules',
           'Topic :: Scientific/Engineering :: Astronomy'],
       platforms=['OS Independent'],
       keywords='meerkat ska',
-      python_requires='>=2.7, !=3.0.*, !=3.1.*, !=3.2.*, <4',
+      python_requires='>=3.5',
       setup_requires=['katversion'],
       use_katversion=True,
       install_requires=['redis>=2.10.5', 'six>=1.12', 'netifaces',
-                        "ipaddress; python_version<'3'", 'msgpack', 'numpy'],
+                        'msgpack', 'numpy'],
       extras_require={'rdb': ['rdbtools', 'python-lzf']},
-      tests_require=['mock', 'rdbtools', 'fakeredis>=1.0.1'])
+      tests_require=['mock', 'rdbtools', 'fakeredis[lua]>=1.3.1'])
