@@ -14,15 +14,12 @@
 # limitations under the License.
 ################################################################################
 
-from __future__ import print_function, division, absolute_import
-from six import raise_from
-
 import logging
 import os.path
 
 from rdbtools import RdbParser, RdbCallback
 
-from .telescope_state import RdbParseError
+from .errors import RdbParseError
 
 
 logger = logging.getLogger(__name__)
@@ -31,7 +28,7 @@ logger = logging.getLogger(__name__)
 class BackendCallback(RdbCallback):
     """A callback adapter that stores keys in backend as RDB file is parsed."""
     def __init__(self):
-        super(BackendCallback, self).__init__(string_escape=None)
+        super().__init__(string_escape=None)
         # Counter keeping track of number of keys inserted into backend
         self.n_keys = 0
         # Flag that helps to disambiguate callback errors from parser errors
@@ -46,7 +43,7 @@ def _parse_rdb_file(parser, callback, fd, filename=None):
         # Don't remap exception to RdbParseError if it originates from callback
         if callback.client_busy:
             raise
-        raise_from(RdbParseError(filename), exc)
+        raise RdbParseError(filename) from exc
 
 
 def load_from_file(callback, file):
